requirements.txt
# Flask and extensions
Flask==2.0.1
Werkzeug==2.3.7

# Database
pymongo==3.12.0

# Cloudflare R2 (S3 Compatible)
boto3==1.28.44

# Environment variables
python-dotenv==0.19.0

# Data processing
pandas==2.1.0
xlsxwriter==3.1.2

# HTTP requests
requests==2.26.0

# Security
bcrypt==4.0.1

# Date and time handling
pytz==2023.3

# File handling
python-magic==0.4.27  # For file type validation

# QR Code generation
qrcode==7.4.2
Pillow==10.0.0  # Required for QR code image generation

# Development tools (optional)
pytest==7.4.2
black==23.7.0
flake8==6.1.0 